fastapi-bg-messaging/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
pillow==10.4.0
pytest==8.2.0
httpx==0.27.2
